Refactor. main파일에 있던 함수들 각 파일에 분리
